更新

+++ build/lib/torchelper/train.py
@@ -5,13 +5,17 @@ from torchelper.models.model_builder import ModelBuilder
 from tqdm import tqdm
 import torch 
 import time
+
 import torch.distributed as dist
 import torch.multiprocessing as mp
-from torchelper.utils.cls_utils import get_cls
+from torchelper.utils.config import merge_cfg
+from torchelper.utils.cls_utils import get_cls, new_cls, auto_new_cls
 from torchelper.data.base_dataset import get_data_loader
 import torch.backends.cudnn as cudnn
 import subprocess
 from torch.utils.tensorboard import SummaryWriter
+from torchelper.utils import logger
+ 
 
 def check_close_port(port):
     result = subprocess.run(['lsof', '-i:'+str(port)], stdout=subprocess.PIPE)
@@ -51,12 +55,13 @@ def check_close_gpu_ids(gpu_ids):
             arr = line.split(' ')
             if len(arr)<8:
                 continue
-            print(line)
             gpu_id = int(arr[1])
             pid = int(arr[4])
             for gid in gpu_ids:
-                if gpu_id == gid:
+                if gpu_id == gid and arr[6].endswith('/python'):
                     os.system('kill '+str(pid))
+                    print('kill', arr[6])
+      
 
 def get_port(def_port):
     result = subprocess.run(['lsof', '-i:'+str(def_port)], stdout=subprocess.PIPE)
@@ -77,6 +82,7 @@ def validate(net:ModelBuilder, epoch:int, val_dataset):
     :raises ValueError: 描述抛出异常场景
     '''
     if get_rank()==0:
+        net.set_eval()
         # net.before_validate(epoch)
         # time.sleep(10)
         
@@ -105,50 +111,76 @@ def validate(net:ModelBuilder, epoch:int, val_dataset):
     torch.distributed.barrier()
 
 #更新tensorboard显示
-def update_vis(dic, tb_writer, epoch, step, step_per_epoch_per_gpu, gpu_count):
-    if dic is None:
+def update_tensorboard(builder:ModelBuilder, tb_writer, epoch, step, step_per_epoch_per_gpu, gpu_count):
+    if tb_writer is None:
+        return
+    audios = builder.get_audio_dict()
+    images = builder.get_img_dict()
+    scalars = builder.get_scalar_dict()
+    metrics = builder.get_metric_dict()
+    audios = audios if audios is not None else {}
+    images = images if images is not None else {}
+    scalars = scalars if scalars is not None else {}
+    metrics = metrics if metrics is not None else {}
+ 
+    step =  (epoch*step_per_epoch_per_gpu+step)*gpu_count
+    # print(epoch, step_per_epoch_per_gpu, epoch*step_per_epoch_per_gpu)
+    for k, v in audios.items():
+        if v is None:
+            logger.warn(k+" is None ... ")
+            continue    
+        tb_writer.add_audio(k, v, step, sample_rate=16000)
+    for k, v in images.items():
+        if v is None:
+            logger.warn(k+" is None ... ")
+            continue
+        tb_writer.add_image(k, v, step)
+    for k, v in scalars.items():
+        if v is None:
+            logger.warn(k+" is None ... ")
+            continue
+        tb_writer.add_scalar(k, v, step)
+    for k, v in metrics.items():
+        if v is None:
+            logger.warn(k+" is None ... ")
+            continue
+        tb_writer.add_scalar(k, v, step)
+
+def update_pbar(builder:ModelBuilder, pbar):
+    if pbar is None:
+        return
+    scalars = builder.get_scalar_dict()
+    if scalars is None:
         return
     msg = []
-    step =  (epoch*step_per_epoch_per_gpu+step)*gpu_count
-    for k, v in dic.items():
-        type = v['type'].lower()
-        val = v['val']
-        if val is None:
-            continue
-        msg.append('%s:%05f' %(k, val))
-        if 'image' in type:
-            tb_writer.add_image(k, val, step)
-        elif 'scalar' in type:
-            tb_writer.add_scalar(k, val, step)
-        elif 'audio' in type:
-            tb_writer.add_audio(k, val, step)
+    for k, v in scalars.items():
+        msg.append('%s:%.5f'%(k, v))
+    pbar.set_description(', '.join(msg))
 
 def train(gpu_id, cfg, is_dist):
-    is_amp = cfg.get('amp', False)
-    train_dataset_cls = get_cls(cfg['train_dataset_cls'])
-    val_dataset_cls = get_cls(cfg['val_dataset_cls'])
-    train_dataset = train_dataset_cls(gpu_id, cfg)
-    train_dataloader = get_data_loader(cfg['batch_per_gpu'], train_dataset, dist=True)
-    val_dataloader = get_data_loader(cfg['batch_per_gpu'], val_dataset_cls(gpu_id, cfg), dist=False)
-    builder:ModelBuilder = get_cls(cfg['model_builder'])(cfg, True, cfg['ckpt_dir'], gpu_id, is_amp)
-    # builder:ModelBuilder = ModelBuilder(True, cfg['ckpt_dir'], gpu_id, is_amp)
+   
+    train_data_cfg = merge_cfg(cfg['dataset']['train'], cfg)
+    val_data_cfg =  merge_cfg(cfg['dataset']['val'], cfg)
+    train_dataset = auto_new_cls(train_data_cfg)
+    val_dataset = auto_new_cls(val_data_cfg)
+    gpu_count = len(cfg['ori_gpu_ids'])
+ 
+    train_dataloader = get_data_loader(cfg['batch_per_gpu'], train_dataset, num_workers=0, dist=True)
+    val_dataloader = get_data_loader(cfg['batch_per_gpu'], val_dataset, dist=False)
+
+    builder:ModelBuilder = get_cls(cfg['model_builder'])(cfg, True, cfg['ckpt_dir'])
+ 
     builder.set_dataset(train_dataset)
     dataset_size = len(train_dataloader)
     
     tb_writer = None
-    gpu_count = len(cfg['ori_gpu_ids'])
     step_per_epoch_per_gpu = dataset_size 
-    if builder.get_vis_dict() is not None and gpu_id==0:
+    if gpu_id==0:
         tb_writer = SummaryWriter(cfg['ckpt_dir'])
 
-    print('#training images = %d' % dataset_size)
-    # save_max_count = cfg.get('save_max_count', -1)
-    # save_max_time = cfg.get('save_max_time', 2*60*60)
-    # if cfg['start_epoch']==0:
-    #     builder.save_model(-1)
     builder.perform_cb('on_begin_train')
     for epoch in range(cfg['start_epoch'], cfg['total_epoch']):
-        # validate(net, epoch, val_dataloader)
+        # validate(builder, epoch, val_dataloader)
         builder.set_train()
         if gpu_id==0:
             pbar = tqdm(train_dataloader)
@@ -170,12 +202,9 @@ def train(gpu_id, cfg, is_dist):
             builder.on_end_backward( epoch, i)
             builder.perform_cb('on_end_step', epoch=epoch, step=i)
             if gpu_id==0:
-                if pbar is not None:
-                    pbar.set_description(builder.get_print_metric())
-                if tb_writer is not None:
-                    update_vis( builder.get_vis_dict(), tb_writer, epoch, i, step_per_epoch_per_gpu, gpu_count)
+                update_pbar(builder, pbar)
+                update_tensorboard(builder, tb_writer, epoch, i, step_per_epoch_per_gpu, gpu_count)
         builder.perform_cb('on_end_epoch', epoch=epoch)
-        builder.update_learning_rate(epoch)
         # builder.save_model(epoch, save_max_count, save_max_time)
         validate(builder, epoch, val_dataloader)
     builder.perform_cb('on_end_train')
